+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - bug fixes
   - bottom horizontal scrollbar not responding on small screens
+- widgets
+  - `logScale` can be set to a number to use a logarithm base different from the default
 
 ## 0.45.5